pyproject.toml
[project]
name = "miniflux-mail-digest"
version = "0.1.0"
description = ""
authors = [{name = "Maor Kadosh", email = "git@avocadosh.xyz"}]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "miniflux == 1.*",
    "beautifulsoup4 == 4.12.*",
    "python-dotenv == 0.21.1", # because of the python version
]

[dependency-groups]
dev = ["types-beautifulsoup4 == 4.12.*"]

[tool.flit.sdist]
include = ["doc/"]

[project.scripts]
miniflux-digest = "miniflux_mail_digest.main:main"

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
preview = true
select = [
    "PIE", "RUF", "F", "E", "W", "I", "UP", "FURB", "ANN", "ASYNC", "B", "A",
    "C4", "T10", "ISC", "INP", "PT", "SIM", "TID", "INT", "PTH", "ERA", "PL",
    "TRY", "PERF", "LOG", "D",
]
ignore = ["D212"]
